refactor: extract coverage analysis into wvImportCoverageController
Move 6 coverage handlers (getCoverage, getCoverageSSE, geoSuggestGap,
dismissCoverageGap, undismissCoverageGap, approveCoverageSuggestion)
plus the fetchGapSubtrees helper out of the
worldViewImportController monolith into a dedicated
wvImportCoverageController file. The monolith re-exports them.

Pure refactor — no behavior change, byte-identical function bodies.
Continues the spine domain-split (ADR-0009).
